chatgpt-api-python/requirements.txt
fastapi
uvicorn

openai
python-dotenv